Added description, ingredients and steps to ice cream recipe
Additions to healthy-chunky-monkey-ice-cream.html include a description header and
a three sentence description. An ingredients header with an unordered
list of ingredients required. A steps header with an ordered list of steps required to
make the dish. Within the steps header, smaller headers were included to further distiguish
between step 1 and 2.